shop_sarafan/requirements.txt
Django==5.0.6
django-debug-toolbar==3.8.1
djangorestframework==3.15.1
djoser==2.2.2
pillow==10.3.0
python-dotenv==1.0.1